stages/gen_3d/pyproject.toml
# NOTE: This stage is Linux + CUDA 12.4 + Python 3.10 only.
# Do not run `uv sync` on macOS — the prebuilt wheels are linux_x86_64.
# Run on the RunPod (or any Linux+A10 box) instead.

[project]
name = "gen-3d"
version = "0.1.0"
description = "Stage 2: image -> 3D model (TRELLIS.2-4B)"
requires-python = ">=3.10,<3.11"
dependencies = [
    "torch==2.6.0",
    "torchvision==0.21.0",
    "triton==3.2.0",
    "setuptools",
    "pillow==12.0.0",
    "matplotlib",
    "rembg",
    "imageio==2.37.2",
    "imageio-ffmpeg==0.6.0",
    "tqdm==4.67.1",
    "easydict==1.13",
    "opencv-python-headless==4.12.0.88",
    "trimesh==4.10.1",
    "zstandard==0.25.0",
    "kornia==0.8.2",
    "timm==1.0.22",
    "transformers==4.57.3",
    "fastapi",
    "uvicorn",
    "diffusers @ git+https://github.com/huggingface/diffusers.git@refs/pull/12790/head",
    "utils3d @ git+https://github.com/EasternJournalist/utils3d.git@9a4eb15e4021b67b12c460c7057d642626897ec8",
    # NOTE: trellis2 itself isn't pip-installable. Clone the repo separately and
    # add it to PYTHONPATH:
    #   git clone https://github.com/microsoft/TRELLIS.2.git /workspace/trellis2-src
    #   export PYTHONPATH=/workspace/trellis2-src:$PYTHONPATH
    "flash-attn-3 @ https://github.com/JeffreyXiang/Storages/releases/download/Space_Wheels_251210/flash_attn_3-3.0.0b1-cp39-abi3-linux_x86_64.whl",
    "cumesh @ https://github.com/JeffreyXiang/Storages/releases/download/Space_Wheels_251210/cumesh-0.0.1-cp310-cp310-linux_x86_64.whl",
    "flex-gemm @ https://github.com/JeffreyXiang/Storages/releases/download/Space_Wheels_251210/flex_gemm-0.0.1-cp310-cp310-linux_x86_64.whl",
    "o-voxel @ https://github.com/JeffreyXiang/Storages/releases/download/Space_Wheels_251210/o_voxel-0.0.1-cp310-cp310-linux_x86_64.whl",
    "nvdiffrast @ https://github.com/JeffreyXiang/Storages/releases/download/Space_Wheels_251210/nvdiffrast-0.4.0-cp310-cp310-linux_x86_64.whl",
    "nvdiffrec-render @ https://github.com/JeffreyXiang/Storages/releases/download/Space_Wheels_251210/nvdiffrec_render-0.0.0-cp310-cp310-linux_x86_64.whl",
]

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu124" }
torchvision = { index = "pytorch-cu124" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/gen_3d"]
